KS-Backend: Azure ADFS Auth backend implementation
The authentication functionality is pretty much handled by the
Azure ADFS library but fetching the user groups from Microsoft
Graph API had to be implemented "manually". The users in HKI city
might have so many groups that they wont all fit in the access
token claims (because of the HTTP header size limitations).

Overwrite the oauth2 callback view to provide custom redirects in
case of success or failure.

Add readme for the usage of the Azure ADFS component.

Refs YJDH-98