Backend: Refactor get_organization_roles
Refactor get_organization_roles so that the roles are saved in
session after login. After login the roles will be fetched from
the session.

This is done because the access token that we get from suomi.fi is
used only for getting these roles and the ttl of the token is only
10 minutes. So later when calling this function we can be assured
that the roles are already fetched and there is no need to call the
suomi.fi organizationRoles API.

Refs YJDH-175