Backend: Refactor authentication logic in shared code
In more detail: There can only be 1 OIDCProfile per user.
This has a bad effect so that if the user logs in the system for
example with 2 different companies on 2 separate computers, they
can only be logged in with one company at a time. In practice the
second company will overwrite the first company. This is a problem
especially in the test environment where the same default user is
used for many tasks and might lead to confusion if multiple users
are using the test environment at the same time with different
companies.

Example of a bad situation with the previous logic:

If a user:

1. Logs in from computer A with Company X
2. Logs in from computer B with Company Y

Computer A company will be switched to Company Y because there can
be only one active session per user. Fix this by linking the active
company to the session instead of the user.

This new logic will use the user session to store the values that
were previously stored in OIDCProfile and EauthorizationProfile.
So these values will be stored in session:

- oidc_id_token
- oidc_access_token
- oidc_refresh_token
- oidc_access_token_expires
- oidc_refresh_token_expires

- eauth_id_token
- eauth_access_token
- eauth_refresh_token (not used atm as suomi.fi does not support
this)
- eauth_access_token_expires
- eauth_refresh_token_expires

Refs YJDH-291