Add /v1/youthapplications/<pk>/activate/ endpoint
The /v1/youthapplications/<pk>/activate/ endpoint can be used to
activate a submitted youth application that hasn't expired. Expiration
is set to 12 hours from submittal of the youth application.

Also send an email with the activation link to the email given in a
submitted youth application after receiving it. Currently the email is
sent to the console when DEBUG=True.

Add tests for activating a youth application:
 - test_youth_applications_activate_invalid_pk
 - test_youth_applications_activate_unexpired_inactive
 - test_youth_applications_activate_unexpired_active
 - test_youth_applications_activate_expired_inactive
 - test_youth_applications_activate_expired_active

Refs YJDH-239, YJDH-334 (Sending an email)
Refs YJDH-348 (Generating the activation link)