KS-Backend: Translate activation email to fi/sv/en
Add langdetect to requirements-dev.(in|txt). Used for detecting the used
language in the activation email's subject and body in the tests.

Revamp the activation email related functions a little bit. Switched to
using gettext here instead of gettext_lazy because without it the email
subject didn't correctly get translated in the tests. Translate the
activation email subject and body according to the youth application's
language.

Add placeholder translations for the activation email's subject and
body for Swedish and English.

Add tests for Finnish, Swedish and English:
 - test_youth_application_post_valid_language
   - Tests that posting a new youth application is successful with
     different languages
 - test_youth_application_activation_email_language
   - Tests that the activation email's subject and body both use the
     youth application's language

Refs YJDH-356