KS-Backend: Compile messages in Docker image build
Compile messages in Docker image build in kesaseteli.Dockerfile to
remove the need for writeable localization directories in the Docker
container. This should fix the error in Platta environment about not
being able to write the generated django.mo files.

Also override the entrypoints in the docker-compose files using the
Kesäseteli backend because although "django-admin compilemessages" is
now run as a part of Docker image build, the resulting files are
overridden by the volume mounts in the docker-compose files and thus
have to regenerated.

Refs YJDH-459

Co-authored-by: Kari Salminen <kari.salminen@anders.com>