YJDH-432: remove rwX access to all files in docker because it might not needed (#691)
This also enables us to remove the copy-on-write layer on container platform so makes sense to try. The way I see it, next.js would only require write if it caches something on disk. Forcing in-memory caching only speeds up the app anyway so it's the logical fix to symptoms caused by this change (if they occur).