KS-Backend: Try different collations in SchoolListView
Previously SchoolListView used "fi-FI-x-icu" collation for sorting the
school list. This failed in the Platta production environment.

Now SchooListView tries out the following collations in order and uses
the first one that works:
 - fi_FI.UTF8
   - PostgreSQL UTF-8 version
 - Finnish_Swedish_CI_AS_UTF8
   - Azure's UTF-8 version
 - fi-FI-x-icu
   - PostgreSQL fallback
 - None
   - No collation override

It was not tested beforehand whether any of the above collations work in
the Platta production environment. If none of the explicit collations
work then the "No collation override" version is used which should
always work. The worst case scenario is that no collation override is
used, the underlying database has not been configured to use
Finnish UTF-8 collation and thus the school list is sorted incorrectly.

Refs YJDH-473