TET: Add ADFS claims givenName and familyName (#899)
TET backend tries to read user's name from ADFS claim. If givenName or familyName is empty, it will produce the following warning but no error: `Claim 'givenName' for user field 'first_name' was not found in the access token for user '6928cf58-4791-480c-b848-1272de20c124'. Field is not required and will be left empty`