KS-Backend: Fix migrations related to target_group
In PR #1234 / commit d4d73280c88f7c1edea0a6593d6362299e83adb9
target_group was renamed but the information added in the migration file
0031_rename_summer_voucher_exception_reason_to_target_group.py
was not completely in sync with the related Django models.

Fix this by renaming the help_text and verbose_name values in the
related Django models to coincide with the manually changed texts in the
migration and make a migration of the remaining incongruence between
EmployerSummerVoucher.target_group and
HistoricalEmployerSummerVoucher.target_group.

Also fix typo PRIMAERY_TARGET_GROUP -> PRIMARY_TARGET_GROUP.

Refs YJDH-631