KS-Backend: Update youth application reactivation tests
Parametrize for all combinations of NEXT_PUBLIC_MOCK_FLAG and
NEXT_PUBLIC_DISABLE_VTJ (i.e. 2 * 2 = 4 combinations):
 - test_youth_applications_activate_unexpired_active
 - test_youth_applications_activate_expired_active

Remove parametrizations for all supported languages from the above to
limit the combinatorial explosion from 48 to 16 test cases per test.
The used youth application factories randomize the language so different
languages are tested although not as thoroughly as with having the
language as a test parameter.

Refs YJDH-647