TET-Backend: Force AMR claim to be a list in ApiTokenAuthentication
Copied FixTunnistamoMixin from palvelutarjotin
https://github.com/City-of-Helsinki/palvelutarjotin/blob/release-v.1.5.0/palvelutarjotin/oidc.py#L4-L19
add renamed it to ForceAmrClaimToListMixin and added reference to the
source.

Make use of ForceAmrClaimToListMixin to fix ApiTokenAuthentication
use with Tunnistamo, where amr claim's not always a list, and newer
authlib which requires that amr claim if truthy is a list. Practically
the same class as KultusApiTokenAuthentication from
https://github.com/City-of-Helsinki/palvelutarjotin/blob/release-v.1.5.0/palvelutarjotin/oidc.py#L40
but simply with different names and docstring.

Refs HL-600