fix(dateinput): clear validation errors when a datepicker selects a date
YJDH-684.

Since the HDS DateInput has a compatibility issue with React-hook-forms,
The onChange should include processes to
1) clear the input validation errors and
2) set the input value.
Otherwise the errors are not cleared when a datepicker is used.

The onChange is currently not in the same format
as the HDS dateinput is expecting it to be.

HDS DateInputProps are waiting for (value: string, valueAsDate: Date),
but onChange from the InputProps is (value: string).
OnChange of the React-hook-Forms waits a SyntheticEvent as a parameter.

NOTE: The HDS v. 3.0.0 will change the DateInput onChange to have an Event as a parameter.