fix: include CSRF token in tet userinfo endpoint response
Currently CSRF token is provided only once during the end of
authentication. If the user loses the CSRF token, there is no way to
retrieve it without logging out and logging back in.

- APIView does not set CSRF token because it defaults to csrf_exempt
- CsrfViewMiddleware does not set a new token if cookie is missing

CSRF token can be forced into the response by calling get_token, so it
is added to the userinfo endpoint to improve the chances of it being
delivered to the browser.

Refs: TETP-306