fix(kesaseteli): accessibility statement link
YJDH-724.

Fix the accessibility statement link in footer of Kesäseteli's Youth and
Employer apps. The link is now targeting to nuorten.hel.fi -page, which
can be dynamically updated.

NOTE: The link is purposely the same in every language and in both the
apps. It was requested so by the product owner.