fix(benefit): issue calculation issue when cloning application
Previously the calculation would not take into account pay subsidies and
training compensations as they were cloned after the calculation.
Judging from the comment "Clone calculation with compensations and pay
subsidies" AND a failing test when application includes pay subsidies or
compensations indicates that the old behaviour may be unintentional.

refs: HL-1625